+++ src/rpc-client/src/AbstractServiceClient.php
@@ -230,14 +230,19 @@ abstract class AbstractServiceClient
         $services = $health->service($this->serviceName)->json();
         $nodes = [];
         foreach ($services as $node) {
-            $passing = false;
+            $passing = true;
             $service = $node['Service'] ?? [];
             $checks = $node['Checks'] ?? [];
 
+            if (isset($service['Meta']['Protocol']) && $this->protocol !== $service['Meta']['Protocol']) {
+                // The node is invalid, if the protocol is not equal with the client's protocol.
+                continue;
+            }
+
             foreach ($checks as $check) {
                 $status = $check['Status'] ?? false;
-                if ($status === 'passing' && $this->protocol === $service['Meta']['Protocol']) {
-                    $passing = true;
+                if ($status !== 'passing') {
+                    $passing = false;
                 }
             }