+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # v1.1.28 - TBD
 
 ## Added
+
 - [#1645](https://github.com/hyperf/hyperf/pull/1645) Added parameter injection support for closure route.
+
+## Fixed
+
+- [#1639](https://github.com/hyperf/hyperf/pull/1639) Fixed bug that the unhealthy node will be got from `consul`.
+
+## Optimized 
+
+- [#1636](https://github.com/hyperf/hyperf/pull/1636) Optimized `co-phpunit` do not broken in coroutine environment, when cases failed.
 
 # v1.1.27 - 2020-04-23